LayoutTests/http/tests/security/xssAuditor/meta-tag-http-refresh-x-frame-options-expected.txt

CONSOLE MESSAGE: Refused to display document because display forbidden by X-Frame-Options.
